Configure UserInterface bean and move all entity validation logic on service layer

+++ src/main/java/org/example/service/impl/AccountServiceImpl.java
@@ -4,6 +4,7 @@ import org.example.dao.AccountDao;
 import org.example.entity.Account;
 import org.example.pojo.CustomerFinancialProfile;
 import org.example.service.AccountService;
+import org.example.service.FinancialProfileService;
 
 import java.util.List;
 
@@ -11,8 +12,11 @@ public class AccountServiceImpl implements AccountService {
 
     private final AccountDao accountDao;
 
-    public AccountServiceImpl(AccountDao accountDao) {
+    private final FinancialProfileService financialProfileService;
+
+    public AccountServiceImpl(AccountDao accountDao, FinancialProfileService financialProfileService) {
         this.accountDao = accountDao;
+        this.financialProfileService = financialProfileService;
     }
 
     @Override
@@ -22,11 +26,21 @@ public class AccountServiceImpl implements AccountService {
 
     @Override
     public void remove(String id) {
+        if (!financialProfileService.existsFinancialProfileById(id)) {
+            throw new IllegalArgumentException("Account with id = %s does not exist!".formatted(id));
+        }
+
         accountDao.removeAccountById(id);
     }
 
     @Override
     public List<CustomerFinancialProfile> getCustomersWithAccountIds() {
-        return accountDao.getCustomersWithAccountIds();
+        List<CustomerFinancialProfile> customersWithAccountIds = accountDao.getCustomersWithAccountIds();
+
+        if (customersWithAccountIds.isEmpty()) {
+            throw new IllegalStateException("No accounts in the system!");
+        }
+
+        return customersWithAccountIds;
     }
 }
